+++ wraptile/src/wraptile/dru_routes.py
@@ -4,7 +4,7 @@ from gavicore.dru_models import OGCApplicationPackage
 from gavicore.dru_service import DRUService
 from gavicore.models import ApiError, ProcessSummary
 
-from .ogcapppkg_response import OgcApplicationPackageResponse
+from .ap_response import OgcApplicationPackageResponse
 from .provider import get_service
 
 dru_router = fastapi.APIRouter()
